+++ src/test/kotlin/racing/CarFactoryTest.kt
@@ -22,15 +22,3 @@ class CarFactoryTest : StringSpec({
         car2.getSequence() shouldBe 1
     }
 })
-
-object CarFactory {
-    private var sequence: Int = 0
-
-    fun createCar(position: Int = 0): Car {
-        return Car.from(sequence++, position)
-    }
-
-    fun resetSequence() {
-        sequence = 0
-    }
-}
